Task 5 review fix: extract shared shift-browse mapper
